--- src/main/java/com/pikachu/purple/application/perfume/service/application/perfume/GetPerfumesByIdsApplicationService.java
@@ -1,23 +1,0 @@
-package com.pikachu.purple.application.perfume.service.application.perfume;
-
-import com.pikachu.purple.application.perfume.port.in.perfume.GetPerfumesByIdsUseCase;
-import com.pikachu.purple.application.perfume.service.domain.PerfumeDomainService;
-import com.pikachu.purple.domain.perfume.Perfume;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-class GetPerfumesByIdsApplicationService implements GetPerfumesByIdsUseCase {
-
-    private final PerfumeDomainService perfumeDomainService;
-
-    @Override
-    public Result invoke(List<Long> perfumeIds) {
-        List<Perfume> perfumes = perfumeDomainService.findAllWithPerfumeAccordsByIds(perfumeIds);
-
-        return new Result(perfumes);
-    }
-
-}
