+++ src/main/java/com/pikachu/purple/application/perfume/service/application/perfume/GetPerfumesService.java
@@ -1,7 +1,6 @@
 package com.pikachu.purple.application.perfume.service.application.perfume;
 
-import com.pikachu.purple.application.perfume.common.dto.PerfumeDTO;
-import com.pikachu.purple.application.perfume.port.in.perfume.GetPerfumesByKeywordUseCase;
+import com.pikachu.purple.application.perfume.port.in.perfume.GetPerfumesUseCase;
 import com.pikachu.purple.application.perfume.service.domain.PerfumeDomainService;
 import com.pikachu.purple.application.review.port.in.starrating.GetAverageScoreByPerfumeIdUseCase;
 import com.pikachu.purple.domain.perfume.Perfume;
@@ -12,11 +11,20 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-class GetPerfumesByKeywordApplicationService implements GetPerfumesByKeywordUseCase {
+class GetPerfumesService implements GetPerfumesUseCase {
 
     private final PerfumeDomainService perfumeDomainService;
     private final GetAverageScoreByPerfumeIdUseCase getAverageScoreByPerfumeIdUseCase;
 
+    private static final int MAX_SIZE = 30;
+
+    @Override
+    public Result invoke(List<Long> perfumeIds) {
+
+        return new Result(perfumeDomainService.findAllWithPerfumeAccordsByIds(perfumeIds));
+    }
+
+    //TODO -> List<PerfumeDTO>
     @Override
     @Transactional
     public Result invoke(String keyword) {
@@ -27,11 +35,20 @@ class GetPerfumesByKeywordApplicationService implements GetPerfumesByKeywordUseC
             perfume.setAverageScore(averageScore);
         }
 
-        List<PerfumeDTO> perfumeDTOs = perfumes.stream()
-            .map(PerfumeDTO::from)
-            .toList();
+        return new Result(perfumes);
+    }
 
-        return new Result(perfumeDTOs);
+    //TODO -> List<RecommendedPerfumeDTO>
+    @Override
+    public Result invoke() {
+        List<Perfume> perfumes =  perfumeDomainService.findAllOrderByReviewCount(MAX_SIZE);
+        for (Perfume perfume : perfumes) {
+            double averageScore = getAverageScoreByPerfumeIdUseCase.invoke(
+                perfume.getId()).averageScore();
+            perfume.setAverageScore(averageScore);
+        }
+
+        return new Result(perfumes);
     }
 
 }
