+++ src/main/java/com/pikachu/purple/bootstrap/perfume/dto/response/GetPerfumesResponse.java
@@ -1,6 +1,6 @@
 package com.pikachu.purple.bootstrap.perfume.dto.response;
 
-import com.pikachu.purple.application.perfume.common.dto.PerfumeDTO;
+import com.pikachu.purple.domain.perfume.Perfume;
 import java.util.List;
 
-public record GetPerfumesResponse(List<PerfumeDTO> perfumes) {}
+public record GetPerfumesResponse(List<Perfume> perfumes) {}
