+++ src/main/java/com/pikachu/purple/bootstrap/recommend/dto/response/GetPerfumesByReviewCountsResponse.java
@@ -1,8 +1,8 @@
 package com.pikachu.purple.bootstrap.recommend.dto.response;
 
-import com.pikachu.purple.application.perfume.common.dto.RecommendedPerfumeDTO;
+import com.pikachu.purple.domain.perfume.Perfume;
 import java.util.List;
 
-public record GetPerfumesByReviewCountsResponse(List<RecommendedPerfumeDTO> perfumes) {
+public record GetPerfumesByReviewCountsResponse(List<Perfume> perfumes) {
 
 }
